refactor: collapse view-settings panel, drop dead ViewSettingsPanel duplicate
